home/requirements.txt
streamlit
Pillow
streamlit_extras
st_pages